# docs/openapi.yaml
openapi: 3.0.3
info:
  title: inLobby Hotel APIs
  version: 0.1.0
servers:
  - url: http://localhost:5000/api/v1
...
paths:
...
  /hotel-orders/orders/hotel:
    get:
      tags: [Orders/Hotel]
      summary: List my hotel orders (office/admin/finance see all)
      operationId: listHotelOrders
      security:
        - XDebugRole: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
          description: Page number (1-based)
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
          description: Items per page
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/HotelOrderPlatformStatus"
          description: Platform status filter (e.g. CONFIRMED, REQUESTED, ...)
        - in: query
          name: rawStatus
          schema:
            $ref: "#/components/schemas/HotelOrderSupplierStatus"
          description: Supplier raw status filter (C/RQ/RJ/RX/X/VCH/VRQ)
        - in: query
          name: role
          schema:
            type: string
            enum: [b2c, b2b_sales_partner, office_user, admin, finance]
          description: Filter by buyer role (office/admin/finance only)
        - in: query
          name: from
          schema: { type: string, format: date-time }
          description: Created-at >= from (ISO)
        - in: query
          name: to
          schema: { type: string, format: date-time }
          description: Created-at <= to (ISO)
        - in: query
          name: q
          schema: { type: string }
          description: Free text search (platformRef, agentRef, email, hotel/city)
      responses:
        "200":
          description: Paged list of orders (fields are role–aware by ACL)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelOrderSummaryPage"
              examples:
                ok:
                  value:
                    page: 1
                    limit: 20
                    total: 2
                    items:
                      - platformRef: "IL-H-20251019-74FD09"
                        status: "CONFIRMED"
                        role: "b2c"
                        summary:
                          agentRef: "inLobby Test"
                          agency: null
                          userEmail: "anna@example.com"
                          city: "Dubai"
                          service: "RAMEE ROSE"
                          rooms: 1
                          leadName: "ARMAN MNATSAKANYAN"
                          arrivalDate: "2025-12-13"
                          nights: 3
                          cancellation: { refundable: true, platformCutoffUtc: "2025-12-06T00:00:00.000Z" }
                          price: { amount: 246.17, currency: "USD" }
                          sellingPrice: { amount: 246.17, currency: "USD" }
                          paid: false
                      - platformRef: "IL-H-20251018-EBB05D"
                        status: "CONFIRMED"
                        role: "b2b_sales_partner"
                        summary:
                          agentRef: "Call-#4382 / Anna"
                          agency: "ACME Travel"
                          userEmail: "agent@acme.tld"
                          city: "Dubai"
                          service: "QUEEN ELIZABETH 2"
                          rooms: 1
                          leadName: "ARMAN MNATSAKANYAN"
                          arrivalDate: "2025-12-13"
                          nights: 3
                          cancellation: { refundable: true, platformCutoffUtc: "2025-12-06T00:00:00.000Z" }
                          price: { amount: 352.26, currency: "USD" }
                          sellingPrice: { amount: 352.26, currency: "USD" }
                          paid: false
        "401":
          description: Unauthorized (login required or X-Debug-Role missing)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden (insufficient privileges)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hotel-orders/orders/hotel/{platformRef}:
    get:
      tags: [Orders/Hotel]
      summary: Get order details by platformRef
      operationId: getHotelOrder
      security:
        - XDebugRole: []
      parameters:
        - in: path
          name: platformRef
          required: true
          schema: { type: string }
          description: The platform reference (e.g. IL-H-YYYYMMDD-XXXXXX)
      responses:
        "200":
          description: Full order details (supplier fields are hidden for non-office roles)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelOrderDetailResponse"
              examples:
                ok:
                  value:
                    order:
                      platformRef: "IL-H-20251019-74FD09"
                      status: "CONFIRMED"
                      role: "b2c"
                      summary:
                        agentRef: "inLobby Test"
                        agency: null
                        userEmail: "anna@example.com"
                        city: "Dubai"
                        service: "RAMEE ROSE"
                        rooms: 1
                        leadName: "ARMAN MNATSAKANYAN"
                        arrivalDate: "2025-12-13"
                        nights: 3
                        cancellation: { refundable: true, platformCutoffUtc: "2025-12-06T00:00:00.000Z" }
                        price: { amount: 246.17, currency: "USD" }
                        sellingPrice: { amount: 246.17, currency: "USD" }
                        paid: false
                      details:
                        hotel:
                          id: "64081"
                          name: "RAMEE ROSE"
                          category: 4
                          address: "Tecom, Dubai"
                          city: "Dubai"
                          country: "United Arab Emirates"
                          image: "https://..."
                        context:
                          hotelSearchCode: "25236127/8511033934022985655/4"
                          arrivalDate: "2025-12-13"
                          nights: 3
                          roomBasis: "Room Only"
                        rooms:
                          - roomId: 1
                            category: "Standard Room"
                            pax:
                              - { type: "ADT", title: "MR.", firstName: "ARMAN", lastName: "MNATSAKANYAN" }
                              - { type: "ADT", title: "MRS.", firstName: "ANNA",  lastName: "GRIGORYAN" }
                        cancellation:
                          platform: { refundable: true, cutoffUtc: "2025-12-06T00:00:00.000Z", bufferDays: 4 }
                          supplier: { refundable: true, deadlineUtc: null }
                          safeToBook: true
                          bufferDays: 4
                        remarksHtml: "<BR />A tax is imposed by the city..."
                        price:
                          retail: { amount: 246.17, currency: "USD" }
                          net: { amount: 239, currency: "USD" }
                      payment:
                        status: "unpaid"
                        method: "none"
                        provider: null
                        history: []
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
  /hotel-orders/orders/hotel/{platformRef}/agent-ref:
    patch:
      tags: [Orders/Hotel]
      summary: Update Agent Reference (internal note field)
      operationId: patchAgentRef
      security:
        - XDebugRole: []
      parameters:
        - in: path
          name: platformRef
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agentRef]
              properties:
                agentRef:
                  type: string
                  maxLength: 120
                  description: Free text internal reference
            examples:
              ex:
                value: { agentRef: "Call-#4382 / Anna" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  agentRef: { type: string }
                required: [ok, agentRef]
              examples:
                ok:
                  value: { ok: true, agentRef: "Call-#4382 / Anna" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
...
components:
  securitySchemes:
    XDebugRole:
      type: apiKey
      in: header
      name: X-Debug-Role
    XDebugMarkup:
      type: apiKey
      in: header
      name: X-Debug-Markup
    XOfferProof:
      type: apiKey
      in: header
      name: X-Offer-Proof
      description: >
        Signed token from Availability (offerProof). If present, Valuation can succeed even when supplier
        omits <Rates/>. Body/query take precedence over this header.
...
  schemas:
    Money:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          description: ISO 4217 (supplier-returned)
          example: EUR
...
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        hint: { type: string }
        context: { type: object }

    # ─────────── Pax / Booking DTOs (kept as-is) ───────────
    PaxPerson:
      type: object
      required: [type, firstName, lastName]
      properties:
        type:
          type: string
          enum: [ADT, CHD]
          description: ADT=adult, CHD=child
        title:
          type: string
          nullable: true
          description: >
            Adults only. Case-sensitive: MR. / MRS. / MISS / MS.
            If omitted or invalid, backend maps to MR.
          enum: ["MR.", "MRS.", "MISS", "MS"]
        firstName:
          type: string
          pattern: "^[A-Za-z][A-Za-z .'-]*$"
          description: Non-unicode only; must start with a letter.
        lastName:
          type: string
          minLength: 2
          pattern: "^[A-Za-z]{2}[A-Za-z .'-]*$"
          description: Non-unicode only; must start with 2 letters.
        age:
          type: integer
          minimum: 0
          maximum: 18
          nullable: true
          description: Required when type=CHD; ignored for ADT.
...
    HotelOrderPlatformStatus:
      type: string
      enum:
        - CONFIRMED
        - REQUESTED
        - REJECTED
        - CANCELLING
        - CANCELLED
        - VOUCHER_ISSUED
        - VOUCHER_REQUESTED
        - PENDING
        - FAILED
    HotelOrderSupplierStatus:
      type: string
      enum: [C, RQ, RJ, RX, X, VCH, VRQ]
    HotelOrderOpsMeta:
      type: object
      properties:
        supplier: { type: string, example: goglobal }
        rawStatus:
          $ref: "#/components/schemas/HotelOrderSupplierStatus"
        rawStatusLabel: { type: string }
        supplierRef: { type: string, nullable: true }
        supplierBookingCode: { type: string, nullable: true }

    # ─────────── NEW: Orders list (flattened items) ───────────
    CancellationSummary:
      type: object
      properties:
        refundable: { type: boolean, nullable: true }
        platformCutoffUtc: { type: string, format: date-time, nullable: true }
    HotelOrderListItem:
      type: object
      properties:
        platformRef: { type: string }
        status:
          $ref: "#/components/schemas/HotelOrderPlatformStatus"
        statusLabel: { type: string }
        isPayable: { type: boolean }
        role: { type: string }
        agentRef: { type: string, nullable: true }
        agency: { type: string, nullable: true }
        user: { type: string, nullable: true }
        city: { type: string, nullable: true }
        service: { type: string, nullable: true }
        rooms: { type: integer }
        leadName: { type: string, nullable: true }
        arrivalDate: { type: string, format: date }
        nights: { type: integer }
        cancellation: { $ref: "#/components/schemas/CancellationSummary" }
        price: { $ref: "#/components/schemas/Money" }
        paid: { type: boolean }
        _ops:
          $ref: "#/components/schemas/HotelOrderOpsMeta"
    HotelOrderSummaryPage:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/HotelOrderListItem" }

    # ─────────── Order details (platform-level) ───────────
...
    PaymentInfo:
      type: object
      properties:
        status:
          type: string
          enum: [unpaid, authorized, captured, paid, settled, refunded, partially_refunded, failed]
        method:
          type: string
          enum: [none, arca, deposit, manual]
          description: Matches order model (none/arca/deposit/manual)
        provider: { type: string, nullable: true }
        history:
          type: array
          items:
            type: object
            properties:
              at: { type: string, format: date-time }
              status: { type: string }
              note: { type: string, nullable: true }
    HotelOrderDetail:
      type: object
      properties:
        platformRef: { type: string }
        status:
          $ref: "#/components/schemas/HotelOrderPlatformStatus"
        statusLabel: { type: string }
        isPayable: { type: boolean }
        role: { type: string }
        supplier:
          type: object
          properties:
            code: { type: string, example: goglobal }
            bookingCode: { type: string, nullable: true }
            supplierRef: { type: string, nullable: true }
            rawStatus:
              $ref: "#/components/schemas/HotelOrderSupplierStatus"
            rawStatusLabel: { type: string }
        hotel:
          type: object
          properties:
            id: { type: string, nullable: true }
            name: { type: string, nullable: true }
            cityName: { type: string, nullable: true }
            countryName: { type: string, nullable: true }
        context:
          type: object
          properties:
            hotelSearchCode: { type: string, nullable: true }
            arrivalDate: { type: string, format: date, nullable: true }
            nights: { type: integer, nullable: true }
            roomsCount: { type: integer, nullable: true }
            roomBasis: { type: string, nullable: true }
        price:
          type: object
          properties:
            amount: { type: number, nullable: true }
            currency: { type: string, nullable: true }
            markupPct: { type: number, nullable: true }
        cancellation:
          $ref: "#/components/schemas/CancellationSummary"
        rooms:
          type: array
          items:
            type: object
            properties:
              roomId: { type: integer }
              pax:
                type: array
                items: { $ref: "#/components/schemas/PaxPerson" }
        supplierRemarksHtml: { type: string, nullable: true }
        clientRemark: { type: string, nullable: true }
        payment: { $ref: "#/components/schemas/PaymentInfo" }
    HotelOrderDetailResponse:
      type: object
      properties:
        order: { $ref: "#/components/schemas/HotelOrderDetail" }
...